requirements.txt
amqp==5.1.1
anyio==3.6.2
billiard==3.6.4.0
celery==5.2.7
cffi==1.15.1
chardet==5.0.0
charset-normalizer==3.0.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==38.0.3
dnspython==2.2.1
fastapi==0.86.0
filetype==1.2.0
h11==0.14.0
idna==3.4
joblib==1.2.0
kombu==5.2.4
motor==3.1.1
numpy==1.23.4
pandas==1.5.1
pdfminer.six==20221105
pdfplumber==0.1.2
prompt-toolkit==3.0.32
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.10.2
# pymongo==4.3.2
PyPDF2==2.11.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.6
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
threadpoolctl==3.1.0
typing_extensions==4.4.0
uvicorn==0.19.0
vine==5.0.0
wcwidth==0.2.5
